+++ src/APRS/APRSTelem.cpp
@@ -142,7 +142,7 @@ uint16_t APRSTelem::decode(uint8_t *data, uint16_t sz)
 
 uint16_t APRSTelem::toJSON(char *json, uint16_t sz, int deviceId)
 {
-    uint16_t result = (uint16_t)snprintf(json, sz, "{\"type\": \"APRSTelem\", \"deviceId\":%d, \"data\": {\"lat\": %.7lf, \"lng\": %.7lf, \"alt\": %lf, \"spd\": %lf, \"hdg\": %lf, \"orient\": [%lf, %lf, %lf], \"stateFlags\": \"%#lx\"}}", deviceId, this->lat, this->lng, this->alt, this->spd, this->hdg, this->orient[0], this->orient[1], this->orient[2], this->stateFlags.get());
+    uint16_t result = (uint16_t)snprintf(json, sz, "{\"type\": \"APRSTelem\", \"deviceId\":%d, \"data\": {\"lat\": %.7lf, \"lng\": %.7lf, \"alt\": %lf, \"spd\": %lf, \"hdg\": %lf, \"orient\": [%lf, %lf, %lf], \"stateFlags\": \"%#lx\"}}", deviceId, this->lat, this->lng, this->alt, this->spd, this->hdg, this->orient[0], this->orient[1], this->orient[2], (uint32_t)this->stateFlags.get());
 
     if (result < sz)
     {
